Fixed our models some
