+++ models/index.js
@@ -1,5 +1,4 @@
 module.exports = {
     Info: require("./sections"),
     Manual: require("./manual"),
-    Company: require("./category")
 };
