--- routes/api/section.js
@@ -1,23 +1,0 @@
-const db = require("../../models");
-const axios = require("axios");
-
-//THIS IS THE CATEGORY
-
-// const manualSchema = new Schema({
-//     Category: {type: String, required: true },
-//     subCategory: [
-//         {
-//             type: Schema.Types.ObjectId,
-//             ref: "Info"
-//         }
-//     ]   
-// })
-
-//CREATE - making sure that the data we're inputting makes it into our schema
-//READ - the data to show in the terminal
-
-
-module.exports = function (app) {
-
-
-};
